// APP/leitura-no-vagao/src/types/Book.tsx
export interface Book {
    ad_livros_id: number;
    titulo: string;
    autor: string;
    editora: string;
    ano_publicacao: string;
    descricao: string;
    isbn10: string;
    isbn13: string;
    criado: string;
    ativo: string;
    categoria_id: number;
    imagem_url: string;
    imagem_base64: string;
}